Move formatters to utility
