+++ tests/unit/utils/number-formatters-test.js
@@ -1,0 +1,10 @@
+import numberFormat from 'labs-nyc-factfinder/utils/number-formatters';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | number format');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = numberFormat();
+  assert.ok(result);
+});
